+++ src/meta/src/controller/catalog/mod.rs
@@ -12,6 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod get_op;
+mod list_op;
+mod test;
+
 use std::collections::{BTreeSet, HashMap, HashSet};
 use std::iter;
 use std::mem::take;
@@ -673,41 +677,6 @@ impl CatalogController {
             .await?;
         txn.commit().await?;
         Ok(())
-    }
-
-    pub async fn list_background_creating_mviews(
-        &self,
-        include_initial: bool,
-    ) -> MetaResult<Vec<table::Model>> {
-        let inner = self.inner.read().await;
-        let status_cond = if include_initial {
-            streaming_job::Column::JobStatus.is_in([JobStatus::Initial, JobStatus::Creating])
-        } else {
-            streaming_job::Column::JobStatus.eq(JobStatus::Creating)
-        };
-        let tables = Table::find()
-            .join(JoinType::LeftJoin, table::Relation::Object1.def())
-            .join(JoinType::LeftJoin, object::Relation::StreamingJob.def())
-            .filter(
-                table::Column::TableType
-                    .eq(TableType::MaterializedView)
-                    .and(
-                        streaming_job::Column::CreateType
-                            .eq(CreateType::Background)
-                            .and(status_cond),
-                    ),
-            )
-            .all(&inner.db)
-            .await?;
-        Ok(tables)
-    }
-
-    pub async fn list_all_object_dependencies(&self) -> MetaResult<Vec<PbObjectDependencies>> {
-        self.list_object_dependencies(true).await
-    }
-
-    pub async fn list_created_object_dependencies(&self) -> MetaResult<Vec<PbObjectDependencies>> {
-        self.list_object_dependencies(false).await
     }
 
     async fn list_object_dependencies(
@@ -1455,16 +1424,6 @@ impl CatalogController {
         Ok(version)
     }
 
-    pub async fn get_secret_by_id(&self, secret_id: SecretId) -> MetaResult<PbSecret> {
-        let inner = self.inner.read().await;
-        let (secret, obj) = Secret::find_by_id(secret_id)
-            .find_also_related(Object)
-            .one(&inner.db)
-            .await?
-            .ok_or_else(|| MetaError::catalog_id_not_found("secret", secret_id))?;
-        Ok(ObjectModel(secret, obj.unwrap()).into())
-    }
-
     pub async fn drop_secret(&self, secret_id: SecretId) -> MetaResult<NotificationVersion> {
         let inner = self.inner.write().await;
         let txn = inner.db.begin().await?;
@@ -1582,20 +1541,6 @@ impl CatalogController {
             )
             .await;
         Ok(version)
-    }
-
-    pub async fn get_connection_by_id(
-        &self,
-        connection_id: ConnectionId,
-    ) -> MetaResult<PbConnection> {
-        let inner = self.inner.read().await;
-        let (conn, obj) = Connection::find_by_id(connection_id)
-            .find_also_related(Object)
-            .one(&inner.db)
-            .await?
-            .ok_or_else(|| MetaError::catalog_id_not_found("connection", connection_id))?;
-
-        Ok(ObjectModel(conn, obj.unwrap()).into())
     }
 
     pub async fn drop_connection(
@@ -2712,44 +2657,6 @@ impl CatalogController {
         Ok(version)
     }
 
-    pub async fn list_stream_job_desc_for_telemetry(
-        &self,
-    ) -> MetaResult<Vec<MetaTelemetryJobDesc>> {
-        let inner = self.inner.read().await;
-        let info: Vec<(TableId, Option<Property>)> = Table::find()
-            .select_only()
-            .column(table::Column::TableId)
-            .column(source::Column::WithProperties)
-            .join(JoinType::LeftJoin, table::Relation::Source.def())
-            .filter(
-                table::Column::TableType
-                    .eq(TableType::Table)
-                    .or(table::Column::TableType.eq(TableType::MaterializedView)),
-            )
-            .into_tuple()
-            .all(&inner.db)
-            .await?;
-
-        Ok(info
-            .into_iter()
-            .map(|(table_id, properties)| {
-                let connector_info = if let Some(inner_props) = properties {
-                    inner_props
-                        .inner_ref()
-                        .get(UPSTREAM_SOURCE_KEY)
-                        .map(|v| v.to_lowercase())
-                } else {
-                    None
-                };
-                MetaTelemetryJobDesc {
-                    table_id,
-                    connector: connector_info,
-                    optimization: vec![],
-                }
-            })
-            .collect())
-    }
-
     pub async fn alter_non_shared_source(
         &self,
         pb_source: PbSource,
@@ -2782,238 +2689,6 @@ impl CatalogController {
             )
             .await;
         Ok(version)
-    }
-
-    pub async fn list_databases(&self) -> MetaResult<Vec<PbDatabase>> {
-        let inner = self.inner.read().await;
-        inner.list_databases().await
-    }
-
-    pub async fn list_schemas(&self) -> MetaResult<Vec<PbSchema>> {
-        let inner = self.inner.read().await;
-        inner.list_schemas().await
-    }
-
-    pub async fn list_all_state_tables(&self) -> MetaResult<Vec<PbTable>> {
-        let inner = self.inner.read().await;
-        inner.list_all_state_tables().await
-    }
-
-    pub async fn list_readonly_table_ids(&self, schema_id: SchemaId) -> MetaResult<Vec<TableId>> {
-        let inner = self.inner.read().await;
-        let table_ids: Vec<TableId> = Table::find()
-            .select_only()
-            .column(table::Column::TableId)
-            .join(JoinType::InnerJoin, table::Relation::Object1.def())
-            .filter(
-                object::Column::SchemaId
-                    .eq(schema_id)
-                    .and(table::Column::TableType.ne(TableType::Table)),
-            )
-            .into_tuple()
-            .all(&inner.db)
-            .await?;
-        Ok(table_ids)
-    }
-
-    pub async fn list_dml_table_ids(&self, schema_id: SchemaId) -> MetaResult<Vec<TableId>> {
-        let inner = self.inner.read().await;
-        let table_ids: Vec<TableId> = Table::find()
-            .select_only()
-            .column(table::Column::TableId)
-            .join(JoinType::InnerJoin, table::Relation::Object1.def())
-            .filter(
-                object::Column::SchemaId
-                    .eq(schema_id)
-                    .and(table::Column::TableType.eq(TableType::Table)),
-            )
-            .into_tuple()
-            .all(&inner.db)
-            .await?;
-        Ok(table_ids)
-    }
-
-    pub async fn list_view_ids(&self, schema_id: SchemaId) -> MetaResult<Vec<ViewId>> {
-        let inner = self.inner.read().await;
-        let view_ids: Vec<ViewId> = View::find()
-            .select_only()
-            .column(view::Column::ViewId)
-            .join(JoinType::InnerJoin, view::Relation::Object.def())
-            .filter(object::Column::SchemaId.eq(schema_id))
-            .into_tuple()
-            .all(&inner.db)
-            .await?;
-        Ok(view_ids)
-    }
-
-    pub async fn list_tables_by_type(&self, table_type: TableType) -> MetaResult<Vec<PbTable>> {
-        let inner = self.inner.read().await;
-        let table_objs = Table::find()
-            .find_also_related(Object)
-            .filter(table::Column::TableType.eq(table_type))
-            .all(&inner.db)
-            .await?;
-        Ok(table_objs
-            .into_iter()
-            .map(|(table, obj)| ObjectModel(table, obj.unwrap()).into())
-            .collect())
-    }
-
-    pub async fn list_sources(&self) -> MetaResult<Vec<PbSource>> {
-        let inner = self.inner.read().await;
-        inner.list_sources().await
-    }
-
-    // Return a hashmap to distinguish whether each source is shared or not.
-    pub async fn list_source_id_with_shared_types(&self) -> MetaResult<HashMap<SourceId, bool>> {
-        let inner = self.inner.read().await;
-        let source_ids: Vec<(SourceId, Option<StreamSourceInfo>)> = Source::find()
-            .select_only()
-            .columns([source::Column::SourceId, source::Column::SourceInfo])
-            .into_tuple()
-            .all(&inner.db)
-            .await?;
-
-        Ok(source_ids
-            .into_iter()
-            .map(|(source_id, info)| {
-                (
-                    source_id,
-                    info.map(|info| info.to_protobuf().cdc_source_job)
-                        .unwrap_or(false),
-                )
-            })
-            .collect())
-    }
-
-    pub async fn list_source_ids(&self, schema_id: SchemaId) -> MetaResult<Vec<SourceId>> {
-        let inner = self.inner.read().await;
-        let source_ids: Vec<SourceId> = Source::find()
-            .select_only()
-            .column(source::Column::SourceId)
-            .join(JoinType::InnerJoin, source::Relation::Object.def())
-            .filter(object::Column::SchemaId.eq(schema_id))
-            .into_tuple()
-            .all(&inner.db)
-            .await?;
-        Ok(source_ids)
-    }
-
-    pub async fn list_sinks(&self) -> MetaResult<Vec<PbSink>> {
-        let inner = self.inner.read().await;
-        inner.list_sinks().await
-    }
-
-    pub async fn list_subscriptions(&self) -> MetaResult<Vec<PbSubscription>> {
-        let inner = self.inner.read().await;
-        inner.list_subscriptions().await
-    }
-
-    pub async fn list_views(&self) -> MetaResult<Vec<PbView>> {
-        let inner = self.inner.read().await;
-        inner.list_views().await
-    }
-
-    pub async fn list_users(&self) -> MetaResult<Vec<PbUserInfo>> {
-        let inner = self.inner.read().await;
-        inner.list_users().await
-    }
-
-    pub async fn get_table_by_name(
-        &self,
-        database_name: &str,
-        table_name: &str,
-    ) -> MetaResult<Option<PbTable>> {
-        let inner = self.inner.read().await;
-        let table_obj = Table::find()
-            .find_also_related(Object)
-            .join(JoinType::InnerJoin, object::Relation::Database2.def())
-            .filter(
-                table::Column::Name
-                    .eq(table_name)
-                    .and(database::Column::Name.eq(database_name)),
-            )
-            .one(&inner.db)
-            .await?;
-        Ok(table_obj.map(|(table, obj)| ObjectModel(table, obj.unwrap()).into()))
-    }
-
-    pub async fn get_table_by_ids(&self, table_ids: Vec<TableId>) -> MetaResult<Vec<PbTable>> {
-        let inner = self.inner.read().await;
-        let table_objs = Table::find()
-            .find_also_related(Object)
-            .filter(table::Column::TableId.is_in(table_ids))
-            .all(&inner.db)
-            .await?;
-        Ok(table_objs
-            .into_iter()
-            .map(|(table, obj)| ObjectModel(table, obj.unwrap()).into())
-            .collect())
-    }
-
-    pub async fn get_sink_by_ids(&self, sink_ids: Vec<SinkId>) -> MetaResult<Vec<PbSink>> {
-        let inner = self.inner.read().await;
-        let sink_objs = Sink::find()
-            .find_also_related(Object)
-            .filter(sink::Column::SinkId.is_in(sink_ids))
-            .all(&inner.db)
-            .await?;
-        Ok(sink_objs
-            .into_iter()
-            .map(|(sink, obj)| ObjectModel(sink, obj.unwrap()).into())
-            .collect())
-    }
-
-    pub async fn get_subscription_by_id(
-        &self,
-        subscription_id: SubscriptionId,
-    ) -> MetaResult<PbSubscription> {
-        let inner = self.inner.read().await;
-        let subscription_objs = Subscription::find()
-            .find_also_related(Object)
-            .filter(subscription::Column::SubscriptionId.eq(subscription_id))
-            .all(&inner.db)
-            .await?;
-        let subscription: PbSubscription = subscription_objs
-            .into_iter()
-            .map(|(subscription, obj)| ObjectModel(subscription, obj.unwrap()).into())
-            .find_or_first(|_| true)
-            .ok_or_else(|| anyhow!("cannot find subscription with id {}", subscription_id))?;
-
-        Ok(subscription)
-    }
-
-    pub async fn get_mv_depended_subscriptions(
-        &self,
-        database_id: Option<DatabaseId>,
-    ) -> MetaResult<HashMap<DatabaseId, HashMap<TableId, HashMap<SubscriptionId, u64>>>> {
-        let inner = self.inner.read().await;
-        let select = Subscription::find()
-            .select_only()
-            .select_column(subscription::Column::SubscriptionId)
-            .select_column(subscription::Column::DependentTableId)
-            .select_column(subscription::Column::RetentionSeconds)
-            .select_column(object::Column::DatabaseId)
-            .join(JoinType::InnerJoin, subscription::Relation::Object.def());
-        let select = if let Some(database_id) = database_id {
-            select.filter(object::Column::DatabaseId.eq(database_id))
-        } else {
-            select
-        };
-        let subscription_objs: Vec<(SubscriptionId, ObjectId, i64, DatabaseId)> =
-            select.into_tuple().all(&inner.db).await?;
-        let mut map: HashMap<_, HashMap<_, HashMap<_, _>>> = HashMap::new();
-        // Write object at the same time we write subscription, so we must be able to get obj
-        for (subscription_id, dependent_table_id, retention_seconds, database_id) in
-            subscription_objs
-        {
-            map.entry(database_id)
-                .or_default()
-                .entry(dependent_table_id)
-                .or_default()
-                .insert(subscription_id, retention_seconds as _);
-        }
-        Ok(map)
     }
 
     pub async fn find_creating_streaming_job_ids(
@@ -3077,155 +2752,6 @@ impl CatalogController {
                 ))
             })
             .collect())
-    }
-
-    pub async fn list_connections(&self) -> MetaResult<Vec<PbConnection>> {
-        let inner = self.inner.read().await;
-        let conn_objs = Connection::find()
-            .find_also_related(Object)
-            .all(&inner.db)
-            .await?;
-        Ok(conn_objs
-            .into_iter()
-            .map(|(conn, obj)| ObjectModel(conn, obj.unwrap()).into())
-            .collect())
-    }
-
-    pub async fn get_all_table_options(&self) -> MetaResult<HashMap<TableId, TableOption>> {
-        let inner = self.inner.read().await;
-        let table_options: Vec<(TableId, Option<i32>)> = Table::find()
-            .select_only()
-            .columns([table::Column::TableId, table::Column::RetentionSeconds])
-            .into_tuple::<(TableId, Option<i32>)>()
-            .all(&inner.db)
-            .await?;
-
-        Ok(table_options
-            .into_iter()
-            .map(|(id, retention_seconds)| {
-                (
-                    id,
-                    TableOption {
-                        retention_seconds: retention_seconds.map(|i| i.try_into().unwrap()),
-                    },
-                )
-            })
-            .collect())
-    }
-
-    pub async fn get_all_created_streaming_parallelisms(
-        &self,
-    ) -> MetaResult<HashMap<ObjectId, StreamingParallelism>> {
-        let inner = self.inner.read().await;
-
-        let job_parallelisms = StreamingJob::find()
-            .filter(streaming_job::Column::JobStatus.eq(JobStatus::Created))
-            .select_only()
-            .columns([
-                streaming_job::Column::JobId,
-                streaming_job::Column::Parallelism,
-            ])
-            .into_tuple::<(ObjectId, StreamingParallelism)>()
-            .all(&inner.db)
-            .await?;
-
-        Ok(job_parallelisms
-            .into_iter()
-            .collect::<HashMap<ObjectId, StreamingParallelism>>())
-    }
-
-    pub async fn get_table_name_type_mapping(
-        &self,
-    ) -> MetaResult<HashMap<TableId, (String, String)>> {
-        let inner = self.inner.read().await;
-        let table_name_types: Vec<(TableId, String, TableType)> = Table::find()
-            .select_only()
-            .columns([
-                table::Column::TableId,
-                table::Column::Name,
-                table::Column::TableType,
-            ])
-            .into_tuple()
-            .all(&inner.db)
-            .await?;
-        Ok(table_name_types
-            .into_iter()
-            .map(|(id, name, table_type)| {
-                (
-                    id,
-                    (name, PbTableType::from(table_type).as_str_name().to_owned()),
-                )
-            })
-            .collect())
-    }
-
-    pub async fn get_table_by_cdc_table_id(
-        &self,
-        cdc_table_id: &String,
-    ) -> MetaResult<Vec<PbTable>> {
-        let inner = self.inner.read().await;
-        let table_objs = Table::find()
-            .find_also_related(Object)
-            .filter(table::Column::CdcTableId.eq(cdc_table_id))
-            .all(&inner.db)
-            .await?;
-        Ok(table_objs
-            .into_iter()
-            .map(|(table, obj)| ObjectModel(table, obj.unwrap()).into())
-            .collect())
-    }
-
-    pub async fn get_created_table_ids(&self) -> MetaResult<Vec<TableId>> {
-        let inner = self.inner.read().await;
-
-        // created table ids.
-        let mut table_ids: Vec<TableId> = StreamingJob::find()
-            .select_only()
-            .column(streaming_job::Column::JobId)
-            .filter(streaming_job::Column::JobStatus.eq(JobStatus::Created))
-            .into_tuple()
-            .all(&inner.db)
-            .await?;
-
-        // internal table ids.
-        let internal_table_ids: Vec<TableId> = Table::find()
-            .select_only()
-            .column(table::Column::TableId)
-            .filter(table::Column::BelongsToJobId.is_in(table_ids.clone()))
-            .into_tuple()
-            .all(&inner.db)
-            .await?;
-        table_ids.extend(internal_table_ids);
-
-        Ok(table_ids)
-    }
-
-    /// Returns column ids of versioned tables.
-    /// Being versioned implies using `ColumnAwareSerde`.
-    pub async fn get_versioned_table_schemas(&self) -> MetaResult<HashMap<TableId, Vec<i32>>> {
-        let res = self
-            .list_all_state_tables()
-            .await?
-            .into_iter()
-            .filter_map(|t| {
-                if t.version.is_some() {
-                    let ret = (
-                        t.id.try_into().unwrap(),
-                        t.columns
-                            .iter()
-                            .map(|c| c.column_desc.as_ref().unwrap().column_id)
-                            .collect_vec(),
-                    );
-                    return Some(ret);
-                }
-                None
-            })
-            .collect();
-        Ok(res)
-    }
-
-    pub async fn list_time_travel_table_ids(&self) -> MetaResult<Vec<TableId>> {
-        self.inner.read().await.list_time_travel_table_ids().await
     }
 }
 
@@ -3622,238 +3148,4 @@ async fn update_internal_tables(
         }
     }
     Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-
-    use risingwave_pb::catalog::StreamSourceInfo;
-
-    use super::*;
-
-    const TEST_DATABASE_ID: DatabaseId = 1;
-    const TEST_SCHEMA_ID: SchemaId = 2;
-    const TEST_OWNER_ID: UserId = 1;
-
-    #[tokio::test]
-    async fn test_database_func() -> MetaResult<()> {
-        let mgr = CatalogController::new(MetaSrvEnv::for_test().await).await?;
-        let pb_database = PbDatabase {
-            name: "db1".to_owned(),
-            owner: TEST_OWNER_ID as _,
-            ..Default::default()
-        };
-        mgr.create_database(pb_database).await?;
-
-        let database_id: DatabaseId = Database::find()
-            .select_only()
-            .column(database::Column::DatabaseId)
-            .filter(database::Column::Name.eq("db1"))
-            .into_tuple()
-            .one(&mgr.inner.read().await.db)
-            .await?
-            .unwrap();
-
-        mgr.alter_name(ObjectType::Database, database_id, "db2")
-            .await?;
-        let database = Database::find_by_id(database_id)
-            .one(&mgr.inner.read().await.db)
-            .await?
-            .unwrap();
-        assert_eq!(database.name, "db2");
-
-        mgr.drop_database(database_id).await?;
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_schema_func() -> MetaResult<()> {
-        let mgr = CatalogController::new(MetaSrvEnv::for_test().await).await?;
-        let pb_schema = PbSchema {
-            database_id: TEST_DATABASE_ID as _,
-            name: "schema1".to_owned(),
-            owner: TEST_OWNER_ID as _,
-            ..Default::default()
-        };
-        mgr.create_schema(pb_schema.clone()).await?;
-        assert!(mgr.create_schema(pb_schema).await.is_err());
-
-        let schema_id: SchemaId = Schema::find()
-            .select_only()
-            .column(schema::Column::SchemaId)
-            .filter(schema::Column::Name.eq("schema1"))
-            .into_tuple()
-            .one(&mgr.inner.read().await.db)
-            .await?
-            .unwrap();
-
-        mgr.alter_name(ObjectType::Schema, schema_id, "schema2")
-            .await?;
-        let schema = Schema::find_by_id(schema_id)
-            .one(&mgr.inner.read().await.db)
-            .await?
-            .unwrap();
-        assert_eq!(schema.name, "schema2");
-        mgr.drop_schema(schema_id, DropMode::Restrict).await?;
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_create_view() -> MetaResult<()> {
-        let mgr = CatalogController::new(MetaSrvEnv::for_test().await).await?;
-        let pb_view = PbView {
-            schema_id: TEST_SCHEMA_ID as _,
-            database_id: TEST_DATABASE_ID as _,
-            name: "view".to_owned(),
-            owner: TEST_OWNER_ID as _,
-            sql: "CREATE VIEW view AS SELECT 1".to_owned(),
-            ..Default::default()
-        };
-        mgr.create_view(pb_view.clone()).await?;
-        assert!(mgr.create_view(pb_view).await.is_err());
-
-        let view = View::find().one(&mgr.inner.read().await.db).await?.unwrap();
-        mgr.drop_relation(ObjectType::View, view.view_id, DropMode::Cascade)
-            .await?;
-        assert!(View::find_by_id(view.view_id)
-            .one(&mgr.inner.read().await.db)
-            .await?
-            .is_none());
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_create_function() -> MetaResult<()> {
-        let mgr = CatalogController::new(MetaSrvEnv::for_test().await).await?;
-        let test_data_type = risingwave_pb::data::DataType {
-            type_name: risingwave_pb::data::data_type::TypeName::Int32 as _,
-            ..Default::default()
-        };
-        let arg_types = vec![test_data_type.clone()];
-        let pb_function = PbFunction {
-            schema_id: TEST_SCHEMA_ID as _,
-            database_id: TEST_DATABASE_ID as _,
-            name: "test_function".to_owned(),
-            owner: TEST_OWNER_ID as _,
-            arg_types,
-            return_type: Some(test_data_type.clone()),
-            language: "python".to_owned(),
-            kind: Some(risingwave_pb::catalog::function::Kind::Scalar(
-                Default::default(),
-            )),
-            ..Default::default()
-        };
-        mgr.create_function(pb_function.clone()).await?;
-        assert!(mgr.create_function(pb_function).await.is_err());
-
-        let function = Function::find()
-            .inner_join(Object)
-            .filter(
-                object::Column::DatabaseId
-                    .eq(TEST_DATABASE_ID)
-                    .and(object::Column::SchemaId.eq(TEST_SCHEMA_ID))
-                    .add(function::Column::Name.eq("test_function")),
-            )
-            .one(&mgr.inner.read().await.db)
-            .await?
-            .unwrap();
-        assert_eq!(function.return_type.to_protobuf(), test_data_type);
-        assert_eq!(function.arg_types.to_protobuf().len(), 1);
-        assert_eq!(function.language, "python");
-
-        mgr.drop_function(function.function_id).await?;
-        assert!(Object::find_by_id(function.function_id)
-            .one(&mgr.inner.read().await.db)
-            .await?
-            .is_none());
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_alter_relation_rename() -> MetaResult<()> {
-        let mgr = CatalogController::new(MetaSrvEnv::for_test().await).await?;
-        let pb_source = PbSource {
-            schema_id: TEST_SCHEMA_ID as _,
-            database_id: TEST_DATABASE_ID as _,
-            name: "s1".to_owned(),
-            owner: TEST_OWNER_ID as _,
-            definition: r#"CREATE SOURCE s1 (v1 int) with (
-  connector = 'kafka',
-  topic = 'kafka_alter',
-  properties.bootstrap.server = 'message_queue:29092',
-  scan.startup.mode = 'earliest'
-) FORMAT PLAIN ENCODE JSON"#
-                .to_owned(),
-            info: Some(StreamSourceInfo {
-                ..Default::default()
-            }),
-            ..Default::default()
-        };
-        mgr.create_source(pb_source).await?;
-        let source_id: SourceId = Source::find()
-            .select_only()
-            .column(source::Column::SourceId)
-            .filter(source::Column::Name.eq("s1"))
-            .into_tuple()
-            .one(&mgr.inner.read().await.db)
-            .await?
-            .unwrap();
-
-        let pb_view = PbView {
-            schema_id: TEST_SCHEMA_ID as _,
-            database_id: TEST_DATABASE_ID as _,
-            name: "view_1".to_owned(),
-            owner: TEST_OWNER_ID as _,
-            sql: "CREATE VIEW view_1 AS SELECT v1 FROM s1".to_owned(),
-            dependent_relations: vec![source_id as _],
-            ..Default::default()
-        };
-        mgr.create_view(pb_view).await?;
-        let view_id: ViewId = View::find()
-            .select_only()
-            .column(view::Column::ViewId)
-            .filter(view::Column::Name.eq("view_1"))
-            .into_tuple()
-            .one(&mgr.inner.read().await.db)
-            .await?
-            .unwrap();
-
-        mgr.alter_name(ObjectType::Source, source_id, "s2").await?;
-        let source = Source::find_by_id(source_id)
-            .one(&mgr.inner.read().await.db)
-            .await?
-            .unwrap();
-        assert_eq!(source.name, "s2");
-        assert_eq!(
-            source.definition,
-            "CREATE SOURCE s2 (v1 INT) WITH (\
-  connector = 'kafka', \
-  topic = 'kafka_alter', \
-  properties.bootstrap.server = 'message_queue:29092', \
-  scan.startup.mode = 'earliest'\
-) FORMAT PLAIN ENCODE JSON"
-        );
-
-        let view = View::find_by_id(view_id)
-            .one(&mgr.inner.read().await.db)
-            .await?
-            .unwrap();
-        assert_eq!(
-            view.definition,
-            "CREATE VIEW view_1 AS SELECT v1 FROM s2 AS s1"
-        );
-
-        mgr.drop_relation(ObjectType::Source, source_id, DropMode::Cascade)
-            .await?;
-        assert!(View::find_by_id(view_id)
-            .one(&mgr.inner.read().await.db)
-            .await?
-            .is_none());
-
-        Ok(())
-    }
 }
